Address Kunal's comments
